refactor(mask): move sprite mask layer to root
