+++ packages/core/src/SpriteMaskLayer.ts
